+++ packages/hostbinder/CHANGELOG.md
@@ -1,6 +1,6 @@
 # @tigojs/hostbinder changelog
 
-## v0.2.21
+## v0.2.21 - v0.2.22
 
 - Minor: upgrade `@tigojs/utils`.