+++ packages/leveldb/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tigojs/leveldb changelog
 
+## v0.2.26
+
+- Minor: upgrade `@tigojs/utils`.
+
+## v0.2.25
+
+- Minor: upgrade `levelup` and `leveldown`.
+
 ## v0.2.23 - v0.2.24
 
 - Minor: upgrade `@tigojs/utils`.