+++ packages/mongodb/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tigojs/mongodb changelog
 
+## v0.3.0
+
+- Breaking change: MongoDB driver has been upgraded to `v4`.
+
+## v0.2.3
+
+- Minor: upgrade `mongodb`.
+
 ## v0.2.1 - v0.2.2
 
 - Minor: upgrade `@tigojs/utils`.